requirements.txt
Flask
werkzeug
pillow
numpy
opencv-python
tensorflow
keras
phi
duckduckgo-search